refactor: split handlers into modules
